requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-ai-projects
azure-identity
azure-cosmos
azure-core
azure-storage-blob
fastapi
uvicorn
pytz
jinja2
python-multipart
aiohttp
opencensus-ext-azure
opencensus-ext-logging
pydantic
pydantic-settings
